Fixed all the warnings and errors.
Release build of game set up for engine proof.
Weapon length is now a float.

+++ MansionMashers/MansionMashers/HeaderFiles/Weapon.h
@@ -24,7 +24,7 @@ typedef struct Weapon
 	int BonusAgility;
 	int BonusDefense;
 
-	int WeaponLength;
+	float WeaponLength;
 	Vec2 WeaponAttackPosition;
 
 	Vec2 Position;
